refactor: JdbcTemplate의 static 메서드들을 목적에 맞게 유틸 클래스로 분리
